added RIO Context & renamed main component file
