+++ client/src/components/RelatedItems/CardCarousel.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect, useContext } from 'react';
 import { ChevronLeft, ChevronRight } from 'react-feather';
+import { RIOContext } from './MainRIO.jsx';
 
 var CardCarousel = (props) => {
   const [scrollAmount, setScroll] = useState(0)
@@ -28,13 +29,16 @@ var CardCarousel = (props) => {
     transform: `translateX(${scrollAmount}px)`
   }
 
+  let state = useContext(RIOContext);
+
+  console.log('--RIO Context--', state)
+
   return (
     <div className="RIC-card-carousel">
       <div className="RIC-arrow-div">{index === 0 ? null : <span onClick={() => leftClick()} ><ChevronLeft className="RIC-left" size={30}/></span>}</div>
       <div className="RIC-carousel-view">
         <div className="RIC-cards" style={scrollCSS}>
-          {/*  conditionally plug in a set of cards to render  */}
-          {props.cards}
+          {/* {props.cards === 'related-items' ? <RelatedItemCards /> : <OutfitsCards />} */}
         </div>
       </div>
       <div className="RIC-arrow-div">{index <= props.data.length - displayLimit ? <span onClick={() => rightClick()} ><ChevronRight className="RIC-right" size={30} /></span> : null}</div>
